Factor out some common test utilities
